backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
anthropic
python-dotenv